app/requirements.txt
wheel
youtube_dl==2021.3.2
pillow==8.1.1
pyyaml==5.4.1
moviepy==1.0.3
ffmpeg==1.4
telethon==1.20
aiosqlite==0.17.0
pyTelegramBotAPI==3.7.6
python-telegram-bot-calendar==1.0.5
apscheduler==3.7.0
tzlocal==2.1
